+++ packages/core/src/generators/angular/index.ts
@@ -25,19 +25,21 @@ import {
 } from '@/helpers/strip-state-and-props-refs';
 import { collectCss } from '@/helpers/styles/collect-css';
 import { nodeHasCss } from '@/helpers/styles/helpers';
+import { traverseNodes } from '@/helpers/traverse-nodes';
 import {
   runPostCodePlugins,
   runPostJsonPlugins,
   runPreCodePlugins,
   runPreJsonPlugins,
 } from '@/modules/plugins';
+import { hashCodeAsString } from '@/symbols/symbol-processor';
 import { MitosisComponent } from '@/types/mitosis-component';
 import { Binding, MitosisNode, checkIsForNode } from '@/types/mitosis-node';
 import { TranspilerGenerator } from '@/types/transpiler';
 import { flow, pipe } from 'fp-ts/lib/function';
 import { isString, kebabCase, uniq } from 'lodash';
+import traverse from 'neotraverse/legacy';
 import { format } from 'prettier/standalone';
-import traverse from 'traverse';
 import isChildren from '../../helpers/is-children';
 import { stringifySingleScopeOnMount } from '../helpers/on-mount';
 import { HELPER_FUNCTIONS, getAppropriateTemplateFunctionKeys } from './helpers';
@@ -49,16 +51,16 @@ import {
 } from './types';
 
 const mappers: {
-  [key: string]: (json: MitosisNode, options: ToAngularOptions) => string;
+  [key: string]: (root: MitosisComponent, json: MitosisNode, options: ToAngularOptions) => string;
 } = {
-  Fragment: (json, options) => {
+  Fragment: (root, json, options) => {
     return `<ng-container>${json.children
-      .map((item) => blockToAngular(item, options))
+      .map((item) => blockToAngular({ root, json: item, options }))
       .join('\n')}</ng-container>`;
   },
-  Slot: (json, options) => {
+  Slot: (root, json, options) => {
     const renderChildren = () =>
-      json.children?.map((item) => blockToAngular(item, options)).join('\n');
+      json.children?.map((item) => blockToAngular({ root, json: item, options })).join('\n');
 
     return `<ng-content ${Object.entries({ ...json.bindings, ...json.properties })
       .map(([binding, value]) => {
@@ -171,7 +173,7 @@ const processCodeBlockInTemplate = (code: string) => {
     let stripped = code.replace('Object.values', '');
     return `useObjectDotValues${stripped}`;
   } else if (code.includes('JSON.stringify')) {
-    let obj = code.match(/JSON.stringify\([^)]*\)/g);
+    let obj = code.match(/JSON.stringify\((.*)\)/);
     return `useJsonStringify(${obj})`;
   } else if (code.includes(' as ')) {
     const asIndex = code.indexOf('as');
@@ -206,13 +208,10 @@ const processEventBinding = (key: string, code: string, nodeName: string, custom
 const stringifyBinding =
   (node: MitosisNode, options: ToAngularOptions, blockOptions: AngularBlockOptions) =>
   ([key, binding]: [string, Binding | undefined]) => {
-    if (binding?.type === 'spread') {
+    if (options.state === 'inline-with-wrappers' && binding?.type === 'spread') {
       return;
     }
-    if (key.startsWith('$')) {
-      return;
-    }
-    if (key === 'key') {
+    if (key.startsWith('$') || key.startsWith('"') || key === 'key') {
       return;
     }
     const keyToUse = BINDINGS_MAPPER[key] || key;
@@ -234,11 +233,13 @@ const stringifyBinding =
       // standard html elements need the attr to satisfy the compiler in many cases: eg: svg elements and [fill]
       return ` [attr.${keyToUse}]="${code}" `;
     } else {
-      return `[${keyToUse}]="${processCodeBlockInTemplate(code)}"`;
+      const codeToUse =
+        options.state === 'inline-with-wrappers' ? processCodeBlockInTemplate(code) : code;
+      return `[${keyToUse}]="${codeToUse}"`;
     }
   };
 
-const handleNgOutletBindings = (node: MitosisNode) => {
+const handleNgOutletBindings = (node: MitosisNode, options: ToAngularOptions) => {
   let allProps = '';
   for (const key in node.properties) {
     if (key.startsWith('$')) {
@@ -258,22 +259,28 @@ const handleNgOutletBindings = (node: MitosisNode) => {
     if (key.startsWith('$')) {
       continue;
     }
-    const { code, arguments: cusArgs = ['event'] } = node.bindings[key]!;
+    let { code, arguments: cusArgs = ['event'] } = node.bindings[key]!;
+    if (options.state === 'class-properties') {
+      code = `this.${code}`;
 
-    if (key.includes('props.')) {
-      allProps += `${key.replace('props.', '')}: ${code}, `;
-    } else if (key.includes('.')) {
-      // TODO: handle arbitrary spread props
-      allProps += `${key.split('.')[1]}: ${code}, `;
-    } else if (key.startsWith('on')) {
+      if (node.bindings[key]?.type === 'spread') {
+        allProps += `...${code}, `;
+        continue;
+      }
+    }
+
+    let keyToUse = key.includes('-') ? `'${key}'` : key;
+    keyToUse = keyToUse.replace('state.', '').replace('props.', '');
+
+    if (key.startsWith('on')) {
       const { event, value } = processEventBinding(key, code, node.name, cusArgs[0]);
       allProps += `on${event.charAt(0).toUpperCase() + event.slice(1)}: ${value.replace(
         /\(.*?\)/g,
         '',
-      )}, `;
+      )}.bind(this), `;
     } else {
-      const codeToUse = processCodeBlockInTemplate(code);
-      const keyToUse = key.includes('-') ? `'${key}'` : key;
+      const codeToUse =
+        options.state === 'inline-with-wrappers' ? processCodeBlockInTemplate(code) : code;
       allProps += `${keyToUse}: ${codeToUse}, `;
     }
   }
@@ -289,17 +296,23 @@ const handleNgOutletBindings = (node: MitosisNode) => {
   return allProps;
 };
 
-export const blockToAngular = (
-  json: MitosisNode,
-  options: ToAngularOptions = {},
-  blockOptions: AngularBlockOptions = {
+export const blockToAngular = ({
+  root,
+  json,
+  options = {},
+  blockOptions = {
     nativeAttributes: [],
   },
-): string => {
+}: {
+  root: MitosisComponent;
+  json: MitosisNode;
+  options?: ToAngularOptions;
+  blockOptions?: AngularBlockOptions;
+}): string => {
   const childComponents = blockOptions?.childComponents || [];
 
   if (mappers[json.name]) {
-    return mappers[json.name](json, options);
+    return mappers[json.name](root, json, options);
   }
 
   if (isChildren({ node: json })) {
@@ -316,9 +329,8 @@ export const blockToAngular = (
     }
 
     if (textCode.includes('JSON.stringify')) {
-      let obj = textCode.replace('JSON.stringify', '');
-      obj = obj.replace(/\(.*?\)/, '');
-      return `{{useJsonStringify${obj}}}`;
+      const obj = textCode.replace(/JSON.stringify\(\s*(\w+)\s*,?.*\)/, '$1');
+      return `{{${obj} | json}}`;
     }
 
     return `{{${textCode}}}`;
@@ -328,24 +340,49 @@ export const blockToAngular = (
 
   if (checkIsForNode(json)) {
     const indexName = json.scope.indexName;
-    str += `<ng-container *ngFor="let ${json.scope.forName} of ${json.bindings.each?.code}${
-      indexName ? `; let ${indexName} = index` : ''
-    }">`;
-    str += json.children.map((item) => blockToAngular(item, options, blockOptions)).join('\n');
+    const forName = json.scope.forName;
+
+    // Check if "key" is present for the first child of the for loop
+    if (json.children[0].bindings && json.children[0].bindings.key?.code) {
+      const fnIndex = (root.meta?._trackByForIndex as number) || 0;
+      const trackByFnName = `trackBy${
+        forName ? forName.charAt(0).toUpperCase() + forName.slice(1) : ''
+      }${fnIndex}`;
+      root.meta._trackByForIndex = fnIndex + 1;
+      let code = json.children[0].bindings.key?.code;
+
+      root.state[trackByFnName] = {
+        code: `${trackByFnName}(${indexName ?? '_'}, ${forName}) { return ${code}; }`,
+        type: 'method',
+      };
+
+      str += `<ng-container *ngFor="let ${forName} of ${json.bindings.each?.code}${
+        indexName ? `; let ${indexName} = index` : ''
+      }; trackBy: ${trackByFnName}">`;
+    } else {
+      str += `<ng-container *ngFor="let ${forName} of ${json.bindings.each?.code}${
+        indexName ? `; let ${indexName} = index` : ''
+      }">`;
+    }
+    str += json.children
+      .map((item) => blockToAngular({ root, json: item, options, blockOptions }))
+      .join('\n');
     str += `</ng-container>`;
   } else if (json.name === 'Show') {
     let condition = json.bindings.when?.code;
-    if (condition?.includes('typeof')) {
+    if (options.state === 'inline-with-wrappers' && condition?.includes('typeof')) {
       let wordAfterTypeof = condition.split('typeof')[1].trim().split(' ')[0];
       condition = condition.replace(`typeof ${wordAfterTypeof}`, `useTypeOf(${wordAfterTypeof})`);
     }
     str += `<ng-container *ngIf="${condition}">`;
-    str += json.children.map((item) => blockToAngular(item, options, blockOptions)).join('\n');
+    str += json.children
+      .map((item) => blockToAngular({ root, json: item, options, blockOptions }))
+      .join('\n');
     str += `</ng-container>`;
     // else condition
     if (isMitosisNode(json.meta?.else)) {
       str += `<ng-container *ngIf="!(${condition})">`;
-      str += blockToAngular(json.meta.else, options, blockOptions);
+      str += blockToAngular({ root, json: json.meta.else, options, blockOptions });
       str += `</ng-container>`;
     }
   } else if (json.name.includes('.')) {
@@ -353,11 +390,39 @@ export const blockToAngular = (
       ? kebabCase(json.name)
       : json.name;
 
-    const allProps = handleNgOutletBindings(json);
+    let allProps = handleNgOutletBindings(json, options);
+    if (options.state === 'class-properties') {
+      const inputsPropsStateName = `mergedInputs_${hashCodeAsString(allProps)}`;
+      root.state[inputsPropsStateName] = {
+        code: 'null',
+        type: 'property',
+      };
+      if (!root.hooks.onInit?.code.includes(inputsPropsStateName)) {
+        if (!root.hooks.onInit) {
+          root.hooks.onInit = { code: '' };
+        }
+        root.hooks.onInit.code += `\nthis.${inputsPropsStateName} = {${allProps}};\n`;
+      }
+      if (
+        root.hooks.onUpdate &&
+        root.hooks.onUpdate.length > 0 &&
+        !root.hooks.onUpdate
+          .map((hook) => hook.code)
+          .join('')
+          .includes(inputsPropsStateName)
+      ) {
+        root.hooks.onUpdate.push({
+          code: `this.${inputsPropsStateName} = {${allProps}}`,
+        });
+      }
+      allProps = `${inputsPropsStateName}`;
+    } else {
+      allProps = `{ ${allProps} }`;
+    }
 
     str += `<ng-container *ngComponentOutlet="
       ${elSelector.replace('state.', '').replace('props.', '')};
-      inputs: { ${allProps} };
+      inputs: ${allProps};
       content: myContent;
       ">  `;
 
@@ -394,7 +459,9 @@ export const blockToAngular = (
     str += '>';
 
     if (json.children) {
-      str += json.children.map((item) => blockToAngular(item, options, blockOptions)).join('\n');
+      str += json.children
+        .map((item) => blockToAngular({ root, json: item, options, blockOptions }))
+        .join('\n');
     }
 
     str += `</${elSelector}>`;
@@ -412,7 +479,7 @@ const traverseToGetAllDynamicComponents = (
   traverse(json).forEach((item) => {
     if (isMitosisNode(item) && item.name.includes('.') && item.name.split('.').length === 2) {
       const children = item.children
-        .map((child) => blockToAngular(child, options, blockOptions))
+        .map((child) => blockToAngular({ root: json, json: child, options, blockOptions }))
         .join('\n');
       dynamicTemplate = `<ng-template #${
         item.name.split('.')[1].toLowerCase() + 'Template'
@@ -450,6 +517,190 @@ const processAngularCode =
       }),
       (newCode) => stripStateAndPropsRefs(newCode, { replaceWith }),
     );
+
+const isASimpleProperty = (code: string) => {
+  const expressions = ['==', '===', '!=', '!==', '<', '>', '<=', '>='];
+  const invalidChars = ['{', '}', '(', ')', 'typeof'];
+
+  return !invalidChars.some((char) => code.includes(char)) && !expressions.includes(code);
+};
+
+const generateNewBindingName = (index: number, name: string) =>
+  `node_${index}_${name.replaceAll('.', '_').replaceAll('-', '_')}`;
+
+const handleBindings = (
+  json: MitosisComponent,
+  item: MitosisNode,
+  index: number,
+  forName?: string,
+  indexName?: string,
+) => {
+  for (const key in item.bindings) {
+    if (
+      key.startsWith('"') ||
+      key.startsWith('$') ||
+      key === 'css' ||
+      key === 'ref' ||
+      isASimpleProperty(item.bindings[key]!.code)
+    ) {
+      continue;
+    }
+
+    const newBindingName = generateNewBindingName(index, item.name);
+
+    if (forName) {
+      if (item.name === 'For') continue;
+      if (key === 'key') continue;
+
+      if (key.startsWith('on')) {
+        const { arguments: cusArgs = ['event'] } = item.bindings[key]!;
+        const eventBindingName = `${generateNewBindingName(index, item.name)}_event`;
+        if (
+          item.bindings[key]?.code.trim().startsWith('{') &&
+          item.bindings[key]?.code.trim().endsWith('}')
+        ) {
+          const forAndIndex = `${forName ? `, ${forName}` : ''}${
+            indexName ? `, ${indexName}` : ''
+          }`;
+          const eventArgs = `${cusArgs.join(', ')}${forAndIndex}`;
+          json.state[eventBindingName] = {
+            code: `(${eventArgs}) => ${item.bindings[key]!.code}`,
+            type: 'function',
+          };
+          item.bindings[key]!.code = `state.${eventBindingName}(${eventArgs})`;
+          json.state[newBindingName] = {
+            code: `(${eventArgs}) => (${item.bindings[key]!.code})`,
+            type: 'function',
+          };
+          item.bindings[key]!.code = `state.${newBindingName}($${eventArgs})`;
+        }
+      } else {
+        json.state[newBindingName] = {
+          code: `(${forName}${indexName ? `, ${indexName}` : ''}) => (${item.bindings[key]!.code})`,
+          type: 'function',
+        };
+        item.bindings[key]!.code = `state.${newBindingName}(${forName}${
+          indexName ? `, ${indexName}` : ''
+        })`;
+      }
+    } else if (item.bindings[key]?.code) {
+      if (item.bindings[key]?.type !== 'spread' && !key.startsWith('on')) {
+        json.state[newBindingName] = { code: 'null', type: 'property' };
+        if (!json.hooks['onInit']?.code) {
+          json.hooks['onInit'] = { code: '' };
+        }
+        json.hooks['onInit'].code += `\nstate.${newBindingName} = ${item.bindings[key]!.code};\n`;
+        json.hooks['onUpdate'] = json.hooks['onUpdate'] || [];
+        json.hooks['onUpdate'].push({
+          code: `state.${newBindingName} = ${item.bindings[key]!.code}`,
+        });
+        item.bindings[key]!.code = `state.${newBindingName}`;
+      } else if (key.startsWith('on')) {
+        const { arguments: cusArgs = ['event'] } = item.bindings[key]!;
+        if (
+          item.bindings[key]?.code.trim().startsWith('{') &&
+          item.bindings[key]?.code.trim().endsWith('}')
+        ) {
+          json.state[newBindingName] = {
+            code: `(${cusArgs.join(', ')}) => ${item.bindings[key]!.code}`,
+            type: 'function',
+          };
+          item.bindings[key]!.code = `state.${newBindingName}(${cusArgs.join(', ')})`;
+        }
+      } else {
+        json.state[newBindingName] = { code: `null`, type: 'property' };
+        if (!json.hooks['onInit']?.code) {
+          json.hooks['onInit'] = { code: '' };
+        }
+        json.hooks['onInit'].code += `\nstate.${newBindingName} = {...(${
+          item.bindings[key]!.code
+        })};\n`;
+        json.hooks['onUpdate'] = json.hooks['onUpdate'] || [];
+        json.hooks['onUpdate'].push({
+          code: `state.${newBindingName} = {...(${item.bindings[key]!.code})}`,
+        });
+        item.bindings[newBindingName] = item.bindings[key];
+        item.bindings[key]!.code = `state.${newBindingName}`;
+        delete item.bindings[key];
+      }
+    }
+    index++;
+  }
+  return index;
+};
+
+const handleProperties = (json: MitosisComponent, item: MitosisNode, index: number) => {
+  for (const key in item.properties) {
+    if (key.startsWith('$') || isASimpleProperty(item.properties[key]!)) {
+      continue;
+    }
+    const newBindingName = generateNewBindingName(index, item.name);
+    json.state[newBindingName] = { code: '`' + `${item.properties[key]}` + '`', type: 'property' };
+    item.bindings[key] = { code: `state.${newBindingName}`, type: 'single' };
+    delete item.properties[key];
+    index++;
+  }
+  return index;
+};
+
+const handleAngularBindings = (
+  json: MitosisComponent,
+  item: MitosisNode,
+  index: number,
+  { forName, indexName }: { forName?: string; indexName?: string } = {},
+): number => {
+  if (isChildren({ node: item })) return index;
+
+  index = handleBindings(json, item, index, forName, indexName);
+  index = handleProperties(json, item, index);
+
+  return index;
+};
+
+const classPropertiesPlugin = () => ({
+  json: {
+    pre: (json: MitosisComponent) => {
+      let lastId = 0;
+      traverseNodes(json, (item) => {
+        if (isMitosisNode(item)) {
+          if (item.name === 'For') {
+            const forName = (item.scope as any).forName;
+            const indexName = (item.scope as any).indexName;
+            traverseNodes(item, (child) => {
+              if (isMitosisNode(child)) {
+                (child as any)._traversed = true;
+                lastId = handleAngularBindings(json, child, lastId, {
+                  forName,
+                  indexName,
+                });
+              }
+            });
+          } else if (!(item as any)._traversed) {
+            lastId = handleAngularBindings(json, item, lastId);
+          }
+        }
+      });
+      return json;
+    },
+  },
+});
+
+// if any state "property" is trying to access state.* or props.*
+// then we need to move them to onInit where they can be accessed
+const transformState = (json: MitosisComponent) => {
+  Object.entries(json.state).forEach(([key, value]) => {
+    if (value?.type === 'property') {
+      if (value.code && (value.code.includes('state.') || value.code.includes('props.'))) {
+        const code = stripStateAndPropsRefs(value.code, { replaceWith: 'this' });
+        json.state[key]!.code = 'null';
+        if (!json.hooks.onInit?.code) {
+          json.hooks.onInit = { code: '' };
+        }
+        json.hooks.onInit.code += `\nthis.${key} = ${code};\n`;
+      }
+    }
+  });
+};
 
 export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
   (userOptions = {}) =>
@@ -518,6 +769,10 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
       }),
     ];
 
+    if (options.state === 'class-properties') {
+      options.plugins.push(classPropertiesPlugin);
+    }
+
     if (options.plugins) {
       json = runPreJsonPlugins({ json, plugins: options.plugins });
     }
@@ -551,7 +806,7 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
       }
       return `public ${variableName} : ${variableType}`;
     });
-    const hasConstructor = Boolean(injectables.length || json.hooks?.onInit);
+    const hasConstructor = Boolean(injectables.length);
 
     const props = getProps(json);
     // prevent jsx props from showing up as @Input
@@ -594,23 +849,43 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
     }
 
     const helperFunctions = new Set<string>();
+
     let template = json.children
       .map((item) => {
-        const tmpl = blockToAngular(item, options, {
-          childComponents,
-          nativeAttributes: useMetadata?.angular?.nativeAttributes ?? [],
+        const tmpl = blockToAngular({
+          root: json,
+          json: item,
+          options,
+          blockOptions: {
+            childComponents,
+            nativeAttributes: useMetadata?.angular?.nativeAttributes ?? [],
+          },
         });
-        getAppropriateTemplateFunctionKeys(tmpl).forEach((key) =>
-          helperFunctions.add(HELPER_FUNCTIONS[key]),
-        );
+        if (options.state === 'inline-with-wrappers') {
+          getAppropriateTemplateFunctionKeys(tmpl).forEach((key) =>
+            helperFunctions.add(HELPER_FUNCTIONS(options.typescript)[key]),
+          );
+        }
         return tmpl;
       })
       .join('\n');
+
     if (options.prettier !== false) {
       template = tryFormat(template, 'html');
     }
 
     stripMetaProperties(json);
+
+    const { components: dynamicComponents, dynamicTemplate } = traverseToGetAllDynamicComponents(
+      json,
+      options,
+      {
+        childComponents,
+        nativeAttributes: useMetadata?.angular?.nativeAttributes ?? [],
+      },
+    );
+
+    transformState(json);
 
     const dataString = getStateObjectStringFromComponent(json, {
       format: 'class',
@@ -622,15 +897,6 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
         stateVars,
       }),
     });
-
-    const { components: dynamicComponents, dynamicTemplate } = traverseToGetAllDynamicComponents(
-      json,
-      options,
-      {
-        childComponents,
-        nativeAttributes: useMetadata?.angular?.nativeAttributes ?? [],
-      },
-    );
 
     const hostDisplayCss = options.visuallyIgnoreHostElement ? ':host { display: contents; }' : '';
     const styles = css.length ? [hostDisplayCss, css].join('\n') : hostDisplayCss;
@@ -708,7 +974,7 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
         .filter((item) => !isSlotProperty(item) && item !== 'children')
         .map((item) => {
           const propType = propsTypeRef ? `${propsTypeRef}["${item}"]` : 'any';
-          let propDeclaration = `@Input() ${item}!: ${propType}`;
+          let propDeclaration = `@Input() ${item}${options.typescript ? `!: ${propType}` : ''}`;
           if (json.defaultProps && json.defaultProps.hasOwnProperty(item)) {
             propDeclaration += ` = defaultProps["${item}"]`;
           }
@@ -719,7 +985,10 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
       ${outputs.join('\n')}
 
       ${Array.from(domRefs)
-        .map((refName) => `@ViewChild('${refName}') ${refName}!: ElementRef`)
+        .map(
+          (refName) =>
+            `@ViewChild('${refName}') ${refName}${options.typescript ? '!: ElementRef' : ''}`,
+        )
         .join('\n')}
       
       ${Array.from(dynamicComponents)
@@ -729,11 +998,11 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
               .split('.')[1]
               .toLowerCase()}Template', { static: true }) ${component
               .split('.')[1]
-              .toLowerCase()}TemplateRef!: TemplateRef<any>`,
+              .toLowerCase()}TemplateRef${options.typescript ? '!: TemplateRef<any>' : ''}`,
         )
         .join('\n')}
 
-      ${dynamicComponents.size ? 'myContent?: any[][];' : ''}
+      ${dynamicComponents.size ? `myContent${options.typescript ? '?: any[][];' : ''}` : ''}
 
       ${dataString}
 
@@ -761,23 +1030,32 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
         !hasConstructor && !dynamicComponents.size
           ? ''
           : `constructor(\n${injectables.join(',\n')}${
-              dynamicComponents.size ? '\nprivate vcRef: ViewContainerRef,\n' : ''
-            }) {
-            ${
-              !json.hooks?.onInit
-                ? ''
-                : `
-              ${json.hooks.onInit?.code}
-              `
-            }
-          }
+              dynamicComponents.size
+                ? `\nprivate vcRef${options.typescript ? ': ViewContainerRef' : ''},\n`
+                : ''
+            }) {}
           `
       }
       ${
-        !json.hooks.onMount.length && !dynamicComponents.size
+        !json.hooks.onMount.length && !dynamicComponents.size && !json.hooks.onInit?.code
           ? ''
           : `ngOnInit() {
-              ${stringifySingleScopeOnMount(json)}
+              ${
+                !json.hooks?.onInit
+                  ? ''
+                  : `
+                    ${json.hooks.onInit?.code}
+                    `
+              }
+              ${
+                json.hooks.onMount.length > 0
+                  ? `
+                    if (typeof window !== 'undefined') {
+                      ${stringifySingleScopeOnMount(json)}
+                    }
+                    `
+                  : ''
+              }
               ${
                 dynamicComponents.size
                   ? `
@@ -798,12 +1076,15 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
       ${
         !json.hooks.onUpdate?.length
           ? ''
-          : `ngAfterContentChecked() {
-              ${json.hooks.onUpdate.reduce((code, hook) => {
-                code += hook.code;
-                return code + '\n';
-              }, '')}
-            }`
+          : `ngOnChanges() {
+              if (typeof window !== 'undefined') {
+                ${json.hooks.onUpdate?.reduce((code, hook) => {
+                  code += hook.code;
+                  return code + '\n';
+                }, '')}
+              }
+            }
+                `
       }
 
       ${