+++ packages/core/src/generators/solid/index.ts
@@ -21,8 +21,8 @@ import { TranspilerGenerator } from '@/types/transpiler';
 import { uniq } from 'fp-ts/lib/Array';
 import * as S from 'fp-ts/string';
 import hash from 'hash-sum';
+import traverse from 'neotraverse/legacy';
 import { format } from 'prettier/standalone';
-import traverse from 'traverse';
 import {
   runPostCodePlugins,
   runPostJsonPlugins,
@@ -64,6 +64,9 @@ function getContextString(component: MitosisComponent, options: ToSolidOptions) 
 const getRefsString = (json: MitosisComponent, options: ToSolidOptions) =>
   Array.from(getRefs(json))
     .map((ref) => {
+      // fix prettier issue when encounter `let props.ref`
+      // Prettier playground: https://prettier.io/playground/#N4Igxg9gdgLgprEAuEAzArlMMCW0AEAsgJ4DCEAtgA7QIwAUVAThFQM5ICGUxAlPsAA6UfPgA2cGPmas2AOhxQq6GACU4qANzDRTSeiYj6O0fgA8i5VPx7UAXmAz2CpSvWoAhAF9RAegB8JqK8wl7CcAAeNExSACYanOhiUiTk1LSwmiAANCCsuNBsyKCcTCwA7gAKpQhFKJxi5ZzERbkARkycYADWkgDKnBRwADKKcMioDWxw7Z09-VRdigDmyDBM6DMg0xQ4axtbkVRwTDhDsA0AKidQpThwdZNi07lsKxIAiugQ8BNTWwArNgRPrvOBfH7jJBPF4gACO33glRY7GQIE4bAAtFA4HB4rEciB1pwcGIVmkKJw0Q0xIS3lBlhIAIIwdY4NoqOCVE6jHF-Z5bAAWMAoYgA6oKcPA2IswHA+rUpTgAG5S4hosBsVogZWbACSUHisD6YFOVBgTMNfRgxAk-NhMmmYs6VDRzAeJ2V41yimmMWRnGWlPtW0WTD9aJtxzYppw5sJzEUMDFOFiMEFyAAHAAGXJ6BE4PQBoNU6H-XIwThtFNpjNIABMuXQ00uVce5ZAcAobTx+OG3GW6EDcAAYhAmJTWStqSoICAvF4gA
+      if (ref.includes('.')) return '';
       const typeParameter = (options.typescript && json['refs'][ref]?.typeParameter) || '';
       return `let ${ref}${typeParameter ? ': ' + typeParameter : ''};`;
     })
@@ -91,7 +94,7 @@ function addProviderComponents(json: MitosisComponent, options: ToSolidOptions) 
 
 const DEFAULT_OPTIONS: ToSolidOptions = {
   state: 'signals',
-  stylesType: 'styled-components',
+  stylesType: 'style-tag',
 };
 
 export const componentToSolid: TranspilerGenerator<Partial<ToSolidOptions>> =
@@ -193,6 +196,7 @@ export const componentToSolid: TranspilerGenerator<Partial<ToSolidOptions>> =
 
       ${getRefsString(json, options)}
       ${getContextString(json, options)}
+      ${json.hooks.onInit?.code ?? ''}
 
       ${json.hooks.onMount.map((hook) => `onMount(() => { ${hook.code} })`).join('\n')}
       ${
@@ -242,8 +246,7 @@ export const componentToSolid: TranspilerGenerator<Partial<ToSolidOptions>> =
           .join('\n')}
         ${
           options.stylesType === 'style-tag' && css && css.trim().length > 4
-            ? // We add the jsx attribute so prettier formats this nicely
-              `<style jsx>{\`${css}\`}</style>`
+            ? `<style>{\`${css}\`}</style>`
             : ''
         }
         ${shouldInjectCustomStyles ? `<style>{\`${json.style}\`}</style>` : ''}