+++ packages/core/src/parsers/jsx/context.ts
@@ -1,5 +1,5 @@
 import * as babel from '@babel/core';
-import traverse from 'traverse';
+import traverse from 'neotraverse/legacy';
 import { createMitosisNode } from '../../helpers/create-mitosis-node';
 import { isMitosisNode } from '../../helpers/is-mitosis-node';
 import { traceReferenceToModulePath } from '../../helpers/trace-reference-to-module-path';